\name{cgInternalClasses}
\docType{class}
\alias{cgInternalClasses}
\alias{characterOrExpression-class}
\alias{characterOrNULL-class}
\alias{dataframeMatrixOrNULL-class}
\alias{dataframeOrNULL-class}
\alias{numericOrNULL-class}
\alias{olsfit-class}
\alias{rrfit-class}
\alias{uvfit-class}
\alias{aftfit-class}
\title{
  cg package Internal Virtual Classes
}
\description{
  cg package Internal Virtual Classes designed for polymorphic slots
}
\details{
  The virtual classes

  \code{characterOrExpression}
  
  \code{characterOrNULL}
  
  \code{dataframeMatrixOrNULL}
  
  \code{dataframeOrNULL}
  
  \code{numericOrNULL}
  
  \code{olsfit}
  
  \code{rrfit}
  
  \code{aftfit}
  
  \code{uvfit}

  are used internally by the cg package, and designed as
  polymorphic slots.
}
\author{
  Bill Pikounis and John Oleynick
}
\note{
  Contact \email{cg@billpikounis.net} for bug reports, questions,
  concerns, and comments. 
}

