\name{cgLineColors}
\docType{data}
\alias{cgLineColors}
\title{Color Choice order for Graph Lines}
\description{
  A vector named \code{cgLineColors} that assigns
  colors to lines for graphing methods in the \pkg{cg} package.
}
\usage{cgLineColors}
\format{A vector containing ten values in this order:
  \code{black blue green red orange brown yellow darkblue darkgreen
    darkgreen}
}
\details{
  This is a package internal convenience, and not intended for
  use or modification by the end user.
  
  If more than ten values are needed then recycling will occur.
}
\note{
  Contact \email{cg@billpikounis.net} for bug reports, questions,
  concerns, and comments.
}
\author{
  Bill Pikounis and John Oleynick
}
\keyword{datasets}