\name{cgValidity}
\alias{cgValidity}
\alias{validAlpha} 
\alias{validPower} 
\alias{validDataFormat}
\alias{validBoolean}
\alias{validCharacter}
\alias{validNumeric}
\alias{validNumericOrCensored}
\alias{validList}
\alias{validAtomicVec}
\alias{validArgMatch}
\alias{validDotsArg} 
\alias{validDotsArgs}
\alias{getDotsArgName}            
\alias{parsePartialName}
\alias{reportInvalidArg}
\alias{validEqualLength}
\alias{validArgMatch}
\alias{validArgDigits}
\alias{validArgModel}
\alias{validCGOneFacGroupColDfr}
\alias{validCGOneFacListedDfr}
\alias{validCensor}
\alias{validZeroScore}
\alias{validAddConstant}
\alias{validAft}
\alias{validFitType}
\alias{validErrorDf}
\alias{validComparisonType}
\alias{validEstimates}
\alias{validGrpNames}
\alias{validN}
\alias{validCutoffWt}
\alias{validDenDf}

\title{
  cg package Validity Checks
}
\description{
  cg package Internal Validity functions not intended for user-level
  calls. These are used on input arguments for methods and functions.
}
\details{

  \code{validAlpha
  validPower
  validDataFormat
  validBoolean
  validCharacter
  validNumeric
  validNumericOrCensored
  validList
  validAtomicVec
  validArgMatch
  validDotsArg
  validDotsArgs
  getDotsArgName
  parsePartialName
  reportInvalidArg
  validEqualLength
  validArgMatch
  validArgDigits
  validArgModel}\cr
  \code{validCGOneFacGroupColDfr
  validCGOneFacListedDfr
  validCensor
  validZeroScore
  validAddConstant
  validAft
  validFitType
  validErrorDf
  validComparisonType
  validEstimates
  validGrpNames
  validN
  validCutoffWt
  validDenDf
}

are used internally by the \pkg{cg} package.
}
\author{
  Bill Pikounis, John Oleynick, Eva Ye
}
\note{
  Contact \email{cg@billpikounis.net} for bug reports, questions,
  concerns, and comments. 
}

